requirements.txt
torch
fastapi==0.95.2
uvicorn[standard]
lanarky
langchain
openai
tiktoken
faiss-cpu
gradio
pinecone-client[grpc]
llama-index
ipython
lanarky
prompt_engineering
